requirements.txt
pystray>=0.19.5
Pillow>=10.2.0
sounddevice>=0.4.6
numpy>=1.24.0
keyboard>=0.13.5
pyperclip>=1.8.2
openai>=1.12.0
python-dotenv>=1.0.0
